habitllm/requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
qdrant_client